test: extract tape white-box suites
